+++ src/components/CMPrefsCompaniesListItem.vue
@@ -1,12 +1,13 @@
 <template>
   <tr :class="{ warning: item.warnings > 0 }" @click="$emit('eventDetails', item)">
     <td>{{ item.id }}</td>
-    <td>{{ item.name }}</td>
-    <td>{{ item.acq_settings_name }}</td>
-    <td>{{ item.sbp_settings_name }}</td>
-    <td>{{ item.ofd_settings_name }}</td>
+    <td class="nobr">{{ item.name }}</td>
+    <!-- <td class="nobr">{{ item.acq_settings_name }}</td>
+    <td class="nobr">{{ item.sbp_settings_name }}</td>
+    <td class="nobr">{{ item.acq_ofd_settings_name }}</td>
+    <td class="nobr">{{ item.sbp_ofd_settings_name }}</td> -->
     <td>{{ item.details }}</td>
-    <!-- <td>{{ item.bank_details }}</td> -->
+    <td>{{ item.bank_details }}</td>
   </tr>
 </template>