+++ src/views/GoodsDetailsView.vue
@@ -46,6 +46,7 @@ import UIOptButtons from "@/components/UIOptButtons.vue";
 const kTABS = [
   { name: "ОСНОВНОЕ", id: "general" },
   { name: "ФОТО", id: "photos" },
+  { name: "ВИДЕО", id: "videos" },
   { name: "ЗАКАЗЫ", id: "orders", access: AccessRightsEnum.OrdersView },
 ];
 
@@ -64,7 +65,7 @@ export default {
   },
 
   setup() {
-    const { view, checkAuthRole } = useView("CarsView");
+    const { view, checkAuthRole } = useView("GoodsDetailsView");
 
     view.title = "Товар";
     view.subTitle = "Детализация";